requirement .txt
python               3.7.8
tensorflow           1.14.0
Keras                2.3.1
scipy                1.5.4
requests             2.25.0
Pillow               8.0.1
numpy                1.16.0
matplotlib           3.3.2
removebg             0.4
face-alignment       1.1.1
opencv-python        4.4.0.46

